separated eligibility section from registration, fixed eligibility spelling on some components
